backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
langchain-classic
langchain-core
langchain-text-splitters
supabase
python-dotenv
pypdf
pandas
unstructured
openpyxl
python-multipart
pydantic
pydantic-settings
psycopg2-binary
google-generativeai
